[refactor] LottoMatch type으로 변경
