+++ src/factories/LottoMatchesFactory.ts
@@ -1,11 +1,11 @@
 import { LOTTERY_AWARD } from "@/constants/lottery";
-import LottoMatch from "@/models/LottoMatch";
+import { LottoMatch } from "@/types/LottoMatch";
 
 export default class LottoMatchesFactory {
   static build() {
     const matchingList = [];
-    for (const { hitCount, price } of LOTTERY_AWARD) {
-      matchingList.push(new LottoMatch(hitCount, price));
+    for (const { numHit, price } of LOTTERY_AWARD) {
+      matchingList.push({ numHit, price, numMatch: 0 } as LottoMatch);
     }
     return matchingList;
   }
