+++ src/types/LottoMatch.ts
@@ -1,0 +1,5 @@
+import { Award } from "./Award";
+
+export type LottoMatch = Award & {
+  numMatch: number;
+};
